+++ crates/adapters/blockchain/examples/node_data_tester.rs
@@ -59,21 +59,20 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let wss_rpc_url = get_env_var("RPC_WSS_URL")?;
     let http_rpc_url = get_env_var("RPC_HTTP_URL")?;
 
-    let dex_pool_filter = DexPoolFilters::new(Some(true));
+    let dex_pool_filter = DexPoolFilters::builder()
+        .remove_pools_with_empty_erc20fields(true)
+        .build();
 
     let client_factory = BlockchainDataClientFactory::new();
-    let client_config = BlockchainDataClientConfig::new(
-        Arc::new(chain.clone()),
-        vec![DexType::UniswapV3],
-        http_rpc_url,
-        None, // RPC requests per second
-        None, // Multicall calls per RPC request
-        Some(wss_rpc_url),
-        true, // Use HyperSync for live data
-        None,
-        Some(dex_pool_filter),
-        Some(PostgresConnectOptions::default()),
-    );
+    let client_config = BlockchainDataClientConfig::builder()
+        .chain(Arc::new(chain.clone()))
+        .dex_ids(vec![DexType::UniswapV3])
+        .http_rpc_url(http_rpc_url)
+        .wss_rpc_url(wss_rpc_url)
+        .use_hypersync_for_live_data(true)
+        .pool_filters(dex_pool_filter)
+        .postgres_cache_database_config(PostgresConnectOptions::default())
+        .build();
 
     let mut node = LiveNode::builder(trader_id, environment)?
         .with_name(node_name)